refactor: move `Project` related `admin` code out of `core/admin.py` to `project/admin.py`
